+++ tests/dummy/app/Http/Controllers/Auth/LoginController.php
@@ -48,11 +48,6 @@ class LoginController extends Controller
     protected $redirectTo = '/home';
 
     /**
-     * @var string
-     */
-    protected $api = 'default';
-
-    /**
      * Create a new controller instance.
      *
      * @return void